Added some robots to Windows build and fixed const/non const char* in src/windows (consequences of changes for GCC 4.3)
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@330 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: 02397ae9325f80fcab3bb8c9705f6c8821d7e7ff
Former-commit-id: 930333bc526ec540d810ac434ac73ac41ab733cd

+++ src/drivers/K1999/K1999.cpp
@@ -542,7 +542,7 @@ void CK1999Data::InitTrack(tTrack* track, void **carParmHandle, tSituation *p)
  for (int Step = 128; (Step /= 2) > 0;)
  {
   OUTPUT("Step = " << Step);
-  for (int i = Iterations * int(sqrt(Step)); --i >= 0;)
+  for (int i = Iterations * int(sqrt((double)Step)); --i >= 0;)
    Smooth(Step);
   Interpolate(Step);
  }
